Update Builder and Create BuilderInfo and AggregateQuery
