+++ src/Query/Builder/Builder.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * Copyright (c) 2021 - Terminus Studio (https://Terminus.Studio)
+ * Copyright (c) 2022 - Terminus Studio (https://Terminus.Studio)
  *
  * ezDB - https://github.com/TerminusStudio/ezDB
  *
@@ -9,26 +9,13 @@
 
 namespace TS\ezDB\Query\Builder;
 
-use TS\ezDB\Connection;
-use TS\ezDB\Connections;
-use TS\ezDB\DB;
-use TS\ezDB\Exceptions\ModelMethodException;
+use Closure;
 use TS\ezDB\Exceptions\QueryException;
-use TS\ezDB\Models\Model;
 use TS\ezDB\Query\Raw;
 
-class Builder
+class Builder extends BuilderInfo implements IBuilder
 {
-
-    /**
-     * @var Connection Instance of the current connection to database
-     */
-    protected $connection;
-
-    /**
-     * @var Model Model class that contains related information of the table being accessed.
-     */
-    protected $model;
+    protected QueryBuilderType $type;
 
     /**
      * @var array[] Contains list of all bindings
@@ -41,186 +28,66 @@ class Builder
         'insert' => [],
         'update' => [],
         'order' => [],
-        'limit' => ['limit' => null, 'offset' => 0],
+        'limit' => null,
+        'offset' => null,
         'aggregate' => [],
         'distinct' => false
     ];
 
     /**
-     * @var string[] Contains list of all allowed operators.
-     */
-    protected $operators = [
-        '=' => '=',
-        '<' => '<',
-        '>' => '>',
-        '<=' => '<=',
-        '>=' => '>=',
-        '<>' => '<>',
-        'LIKE' => 'LIKE'
-    ];
-
-    /**
-     * @var array Contains a list of relationships that needs to be eager loaded
-     */
-    protected $eagerLoad = [];
-
-    /**
-     * Builder constructor.
-     * @param Connection|null $connection
-     */
-    public function __construct(Connection $connection = null)
-    {
-        $this->connection = $connection;
-    }
-
-    /**
-     * set Model class
-     * @param Model $model
-     */
-    public function setModel(Model $model = null)
-    {
-        $this->model = $model;
-        if ($model != null) {
-            $this->table($model->getTable());
-        }
-        return $this;
-    }
-
-    /**
-     * Check if the model class is set
-     * @return bool
-     */
-    public function hasModel()
-    {
-        return ($this->model != null);
-    }
-
-    /**
-     * Get current connection
-     * @return Connection
-     * @throws \TS\ezDB\Exceptions\ConnectionException
-     */
-    public function getConnection()
-    {
-        if ($this->connection == null) {
-            $this->connection = Connections::connection();
-        }
-        return $this->connection;
-    }
-
-    /**
-     * This function adds to a binding.
-     * @param $binding
-     * @param string $type
-     */
-    public function addBinding($binding, $type = 'where')
-    {
-        $this->bindings[$type][] = $binding;
-        return $this;
-    }
-
-    /**
-     * This function sets a binding. Overwrites any previous bindings completely.
-     * @param $binding
-     * @param string $type
-     */
-    protected function setBinding($binding, $type = 'where')
-    {
-        $this->bindings[$type] = $binding;
-        return $this;
-    }
-
-    /**
-     * @param string $type
+     * @param $type
      * @return array
+     * @deprecated use getClauses()
      */
-    public function getBindings($type = 'where')
+    public function getBindings($type = 'where'): array
     {
-        return $this->bindings[$type];
+        return $this->getClauses($type);
     }
 
     /**
-     * @param string $type
-     * @return array
-     * @throws \TS\ezDB\Exceptions\ConnectionException
+     * @inheritDoc
      */
-    public function prepareBindings($type = 'select')
+    public function from(string $tableName): static
     {
-        return $this->getConnection()->getDriver()->getProcessor()->$type($this->bindings);
-    }
-
-    /**
-     * @param $table
-     * @return $this
-     */
-    public function table($table)
-    {
-        $this->addBinding($table, 'from');
+        $this->addClause('from', $tableName);
         return $this;
     }
 
     /**
-     * @param $table
-     * @return $this
+     * @inheritDoc
      */
-    public function from($table)
+    public function table(string $tableName): static
     {
-        $this->addBinding($table, 'from');
-        return $this;
+        return static::from($tableName);
     }
 
-
     /**
-     * This function accepts 1d and 2d arrays to insert records.
-     * 1D Array : ['name' => 'John', 'age' => 21];
-     * 2D Array : [ 0 => ['name' => 'John', 'age' => 21], 1=> ['name' => 'Jane', 'age' => 22] ];
-     * Calling this function through the model will update created_at and updated_at timestamps.
-     * @param array $values
-     * @return bool true/false
-     * @throws QueryException
-     * @throws \TS\ezDB\Exceptions\ConnectionException
+     * @inheritDoc
      */
-    public function insert($values)
+    public function insert(array $values): static
     {
+        $this->setType(QueryBuilderType::Insert);
         if (!is_array($values)) {
             throw new QueryException('Invalid insert argument');
         }
 
         if (is_array(current($values))) {
             foreach ($values as $value) {
-                if ($this->hasModel() && $this->model->hasTimestamps()) {
-                    $value[$this->model->getCreatedAt()] = $this->now();
-                    $value[$this->model->getUpdatedAt()] = $this->now();
-                }
-
                 ksort($value);
-                //TODO: Check if all the arrays contain the same key
-
-                $this->addBinding($value, 'insert');
+                static::insert($values);
             }
         } else {
-            if ($this->hasModel() && $this->model->hasTimestamps()) {
-                $values[$this->model->getCreatedAt()] = $this->now();
-                $values[$this->model->getUpdatedAt()] = $this->now();
-            }
-
-            $this->addBinding($values, 'insert');
+            $this->addClause('insert', $values);
         }
-
-        [$sql, $params] = $this->prepareBindings('insert');
-
-        return $this->connection->insert($sql, ...$params);
+        return $this;
     }
 
     /**
-     * Update a column with a given value
-     * Update values can either be called using set method or passing a array to this method
-     * @param array $values Accepts any key value array
-     * @return array|bool|mixed
-     * @throws \TS\ezDB\Exceptions\ConnectionException
+     * @inheritDoc
      */
-    public function update($values = null)
+    public function update(?array $values = null): static
     {
+        $this->setType(QueryBuilderType::Update);
         if ($values != null) {
             if (!is_array($values)) {
                 throw new QueryException('Invalid update arguments');
@@ -230,508 +97,291 @@ class Builder
                 $this->set($column, $value);
             }
         }
-
-        if ($this->hasModel() && $this->model->hasTimestamps()) {
-            $this->set($this->model->getUpdatedAt(), $this->now());
-        }
-
-        [$sql, $params] = $this->prepareBindings('update');
-
-        return $this->connection->update($sql, ...$params);
-    }
-
-    /**
-     * @param $table
-     * @param $condition1
-     * @param null $operator
-     * @param null $condition2
-     * @param string $joinType
-     * @return $this
-     * @throws QueryException
-     */
-    public function join(
-        $table,
-        $condition1,
-        $operator = null,
-        $condition2 = null,
-        $joinType = 'INNER JOIN'
-    )
-    {
-        if ($condition1 instanceof \Closure) {
-            $type = 'nested';
-            $condition1($query = new JoinBuilder($this)); //call the function with new instance of join builder
-            $nested = $query->getBindings();
-            if ($operator != null) {
-                //Assume operator is the joinType for nested statements
-                $joinType = $operator;
-            }
-            $this->addBinding(compact('table', 'nested', 'joinType', 'type'), 'join');
-            return $this;
-        }
-        if ($condition2 == null) {
-            throw new QueryException('Invalid Condition');
-        }
-        if ($this->isInvalidOperator($operator)) {
-            throw new QueryException('Invalid Operator');
-        }
-
-        $type = 'basic';
-        $this->addBinding(
-            compact('table', 'condition1', 'operator', 'condition2', 'joinType', 'type'),
-            'join'
-        );
         return $this;
     }
 
     /**
-     * @param string|\Closure|Raw $column
-     * @param string|array|null $operator
-     * @param string|null $value
-     * @param string $boolean
-     * @return $this
-     * @throws QueryException
+     * @inheritDoc
      */
-    public function where($column, $operator = null, $value = null, $boolean = 'AND')
+    public function select(array|string $columns = ['*']): static
+    {
+        $this->setType(QueryBuilderType::Select);
+        if (!is_array($columns)) {
+            $columns = [$columns];
+        }
+
+        foreach ($columns as $column) {
+            $this->addClause('select', $column);
+        }
+
+        return $this;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function get(array|string $columns = ['*']): static
+    {
+        return $this->select($columns);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function first(array|string $columns = ['*']): static
+    {
+        return $this->limit(1)->select($columns);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function delete(): static
+    {
+        $this->setType(QueryBuilderType::Delete);
+        return $this;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function truncate(): static
+    {
+        $this->setType(QueryBuilderType::Truncate);
+        return $this;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function set(string $column, object $value): static
+    {
+        $this->addClause('update', ['column' => $column, 'value' => $value]);
+        return $this;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function where(string|Closure|array $column, ?string $operator = null, ?object $value = null, string $boolean = 'AND'): static
     {
         if (is_array($column)) {
-            foreach ($column as $value) {
-                if (!is_array($value)) {
+            foreach ($column as $whereCondition) {
+                if (!is_array($whereCondition)) {
                     throw new QueryException('Invalid Array of Values');
                 }
-                $this->where(...array_values($value));
+                $this->where(...array_values($whereCondition));
             }
             return $this;
         } elseif ($column instanceof \Closure) {
             $type = 'nested';
             $column($query = new static()); //call the function with new static instance
-            $nested = $query->getBindings('where'); //get bindings
-            $this->addBinding(compact('nested', 'boolean', 'type'), 'where');
+            $nested = $query->getClauses('where');
+            $this->addClause('where', ['nested' => $nested, 'boolean' => $boolean, 'type' => $type]);
             return $this;
-        } elseif ($column instanceof Raw) {
-            $operator = $operator ?? []; //Oprator contains bindings.
-            return $this->whereRaw($column, $operator, $boolean);
         }
+
         if (is_null($value)) {
             if (is_null($operator)) {
                 throw new QueryException('Null Operator and Value. Did you mean to call whereNull()');
             }
 
             $value = $operator;
-            $operator = "=";
+            $operator = '=';
         } elseif ($this->isInvalidOperator($operator)) {
             throw new QueryException('Invalid Operator');
         }
+
         $type = 'basic';
-        $this->addBinding(compact('column', 'operator', 'value', 'boolean', 'type'), 'where');
+        $this->addClause('where', ['column' => $column, 'operator' => $operator, 'value' => $value, 'boolean' => $boolean, 'type' => $type]);
         return $this;
     }
 
     /**
-     * @param $column
-     * @param null $operator
-     * @param null $value
-     * @return $this
-     * @throws QueryException
+     * @inheritDoc
      */
-    public function orWhere($column, $operator = null, $value = null)
+    public function orWhere(string|Closure|array $column, ?string $operator = null, ?object $value = null): static
     {
         return $this->where($column, $operator, $value, 'OR');
     }
 
     /**
-     * @param $column
-     * @param string $boolean
-     * @param false $not
-     * @return $this
+     * @inheritDoc
      */
-    public function whereNull($column, $boolean = 'AND', $not = false)
+    public function whereNull(string $column, string $boolean = 'AND', bool $not = false): static
     {
-        $type = 'isNull';
-        if (is_array($column)) {
-            foreach ($column as $c) {
-                return $this->whereNull($c, $boolean, $not);
-            }
-        }
-        $this->addBinding(compact('column', 'type', 'boolean', 'not'), 'where');
+        $this->addClause('where', [
+            'type' => 'null',
+            'column' => $column,
+            'boolean' => $boolean,
+            'not' => $not
+        ]);
         return $this;
     }
 
     /**
-     * @param $column
-     * @param string $boolean
-     * @return $this
+     * @inheritDoc
      */
-    public function whereNotNull($column, $boolean = 'AND')
+    public function whereNotNull(string $column, string $boolean = 'AND'): static
     {
         return $this->whereNull($column, $boolean, true);
     }
 
     /**
-     * @param $column
-     * @param array $value
-     * @param string $boolean
-     * @param false $not
-     * @return $this
+     * @inheritDoc
      */
-    public function whereBetween($column, array $value, $boolean = 'AND', $not = false)
+    public function whereBetween(string $column, object $value1, object $value2, string $boolean = 'AND', bool $not = false): static
     {
-        $type = 'between';
-        /*if (is_array($column)) {
-            foreach ($column as $c) {
-                return $this->whereBetween($c, $value, $boolean, $not);
-            }
-        }*/
-
-        if (count($value) !== 2) {
-            throw new QueryException("Value array length must be 2.");
-        }
-
-        $this->addBinding(compact('column', 'type', 'value', 'boolean', 'not'), 'where');
+        $this->addClause('where', [
+            'type' => 'between',
+            'column' => $column,
+            'value1' => $value1,
+            'value2' => $value2,
+            'boolean' => $boolean,
+            'not' => $not
+        ]);
         return $this;
     }
 
     /**
-     * @param $column
-     * @param array|null $value
-     * @param string $boolean
-     * @return $this
+     * @inheritDoc
      */
-    public function whereNotBetween($column, array $value = null, $boolean = 'AND')
+    public function whereNotBetween(string $column, object $value1, object $value2, string $boolean = 'AND'): static
     {
-        return $this->whereBetween($column, $value, $boolean, true);
+        return $this->whereBetween($column, $value1, $value2, $boolean, true);
     }
 
     /**
-     * @param string $column
-     * @param array $values
-     * @param string $boolean
-     * @param false $not
-     * @return $this
+     * @inheritDoc
      */
-    public function whereIn($column, $values, $boolean = 'AND', $not = false)
+    public function whereIn(string $column, array $values, string $boolean = 'AND', bool $not = false): static
     {
-        $type = 'in';
-
-        $this->addBinding(compact('column', 'type', 'values', 'boolean', 'not'), 'where');
-
+        $this->addClause('where', [
+            'type' => 'in',
+            'column' => $column,
+            'values' => $values,
+            'boolean' => $boolean,
+            'not' => $not
+        ]);
         return $this;
     }
 
     /**
-     * @param $column
-     * @param $values
-     * @param string $boolean
-     * @return $this
+     * @inheritDoc
      */
-    public function whereNotIn($column, $values, $boolean = 'AND')
+    public function whereNotIn(string $column, array $values, string $boolean = 'AND'): static
     {
         return $this->whereIn($column, $values, $boolean, true);
     }
 
     /**
-     * Execute raw where statements.
-     *
-     * @param string|Raw $raw
-     * @param array $values
-     * @param string $boolean
-     * @return $this
-     * @throws QueryException
+     * @inheritDoc
      */
-    public function whereRaw($raw, $values = [], $boolean = 'AND')
+    public function whereRaw(string|Raw $raw, string $boolean = 'AND'): static
     {
-        $type = 'raw';
-        $values = (array)$values;
-
         if (is_string($raw)) {
             $raw = new Raw($raw);
         } elseif (!$raw instanceof Raw) {
             throw new QueryException('$raw must be an instance of Raw class or a string,');
         }
-
-        $this->addBinding(compact('raw', 'values', 'boolean', 'type'), 'where');
-
+        $this->addClause('where', [
+            'type' => 'in',
+            'raw' => $raw,
+            'boolean' => $boolean
+        ]);
         return $this;
     }
 
     /**
-     * @param $column
-     * @param string $direction
-     * @throws QueryException
+     * @inheritDoc
      */
-    public function orderBy($column, $direction = 'asc')
+    public function orderBy(string $column, string $direction = 'ASC'): static
     {
-        $direction = strtolower($direction);
-        if ($direction !== 'asc' && $direction !== 'desc') {
+        $direction = strtoupper($direction);
+        if ($direction !== 'ASC' && $direction !== 'DESC') {
             throw new QueryException("Order Direction must be ASC or DESC");
         }
 
-        $this->addBinding(compact('column', 'direction'), 'order');
+        $this->addClause('order', ['column' => $column, 'direction' => $direction]);
         return $this;
     }
 
     /**
-     * @param $limit
-     * @param null $offset
-     * @return $this
+     * @inheritDoc
      */
-    public function limit($limit, $offset = null)
+    public function limit(int $limit, ?int $offset = null): static
     {
-        $this->bindings['limit']['limit'] = $limit;
-        if ($offset !== null) {
-            return $this->offset($offset);
-        }
+        $this->addClause('limit', $limit, replace: true);
+        if ($offset != null)
+            $this->addClause('offset', $offset, replace: true);
         return $this;
     }
 
     /**
-     * @param $offset
-     * @return $this
+     * @inheritDoc
      */
-    public function offset($offset)
+    public function offset(int $offset): static
     {
-        $this->bindings['limit']['offset'] = $offset;
+        $this->addClause('offset', $offset, replace: true);
         return $this;
     }
 
     /**
-     * This method is used for the update. Each column and value can be set separately.
-     * Use update method itself for setting using arrays
-     *
-     * @param $column
-     * @param $value
-     * @return $this
+     * @inheritDoc
      */
-    public function set($column, $value)
-    {
-        //TODO: Maybe check whether the column was already set?
-        $this->addBinding(compact('column', 'value'), 'update');
-        return $this;
-    }
-
-    /**
-     * This function should be used with the model for eager loading.
-     *
-     * TODO: Support loading relations in a single query.
-     * SELECT users.*, '' as `with`, api.user_id is null as `exists`, api.* FROM users
-     * LEFT JOIN api ON users.id = api.user_id
-     *
-     * @param string|array $relations
-     * @return Builder
-     * @throws ModelMethodException
-     */
-    public function with($relations)
-    {
-        if (!$this->hasModel()) {
-            throw new ModelMethodException("with() method is only accessible when using Models.");
-        }
-
-        if (is_array($relations)) {
-            $this->eagerLoad = array_merge($this->eagerLoad, $relations);
-        } else {
-            $this->eagerLoad[] = $relations;
-        }
-
-        return $this;
-    }
-
-    public function distinct($set = true)
-    {
-        $this->bindings['distinct'] = $set;
-        return $this;
-    }
-
-    /**
-     * @param string|string[] $columns
-     * @return array|bool|mixed
-     * @throws \TS\ezDB\Exceptions\ConnectionException|\TS\ezDB\Exceptions\QueryException
-     * @throws ModelMethodException
-     */
-    public function get($columns = ['*'])
+    public function count(array|string $columns = ['*']): IAggregateQuery
     {
         if (!is_array($columns)) {
             $columns = [$columns];
         }
-
-        foreach ($columns as $column) {
-            $this->addBinding($column, 'select');
-        }
-
-        [$sql, $params] = $this->prepareBindings('select');
-
-        $r = $this->connection->select($sql, ...$params);
-
-        if (!$this->hasModel()) {
-            return $r;
-        }
-
-        return $this->model::createFromResult($r, $this->eagerLoad);
+        return $this->aggregate('count', $columns);
     }
 
     /**
-     * Delete from table. Only works if conditions are set, to delete all rows use truncate().
-     *
-     * @return array|bool|int|mixed
-     * @throws QueryException
-     * @throws \TS\ezDB\Exceptions\ConnectionException|\TS\ezDB\Exceptions\QueryException
+     * @inheritDoc
      */
-    public function delete()
-    {
-        //For safety make sure there is some conditions set.
-        //TODO: Maybe let user specify a force delete all behaviour
-        if (empty($this->getBindings('where'))) {
-            throw new QueryException(
-                'delete() method was called without any conditions. ' .
-                'If you want to delete all the rows, use truncate() instead.'
-            );
-        }
-
-        [$sql, $params] = $this->prepareBindings('delete');
-
-        return $this->connection->delete($sql, ...$params);
-    }
-
-    /**
-     * Truncate a table.
-     *
-     * @return array|bool|int|mixed|object
-     * @throws \TS\ezDB\Exceptions\ConnectionException|\TS\ezDB\Exceptions\QueryException
-     */
-    public function truncate()
-    {
-        $sql = $this->prepareBindings('truncate');
-
-        $result = $this->connection->raw($sql);
-        return ($result || empty($result));
-    }
-
-    /**
-     * Get the first row of result
-     *
-     * @param string[] $columns
-     * @return array|bool|mixed
-     * @throws \TS\ezDB\Exceptions\ConnectionException
-     */
-    public function first($columns = ['*'])
-    {
-        $this->limit(1, 0);
-        $r = $this->get($columns);
-        //Select the first object from the array and return.
-        return $r[0] ?? $r;
-    }
-
-    /**
-     * Get the count result of the query.
-     *
-     * @param string $columns
-     * @return mixed
-     * @throws QueryException
-     * @throws \TS\ezDB\Exceptions\ConnectionException
-     */
-    public function count($columns = '*')
-    {
-        if (!is_array($columns)) {
-            $columns = [$columns];
-        }
-
-        return intval($this->aggregate('count', $columns));
-    }
-
-    /**
-     * Find the sum of the given column.
-     *
-     * @param $column
-     * @return mixed
-     * @throws QueryException
-     * @throws \TS\ezDB\Exceptions\ConnectionException
-     */
-    public function sum($column)
+    public function sum(string $column)
     {
         return $this->aggregate('sum', [$column]);
     }
 
     /**
-     * Find the avg of a given column.
-     *
-     * @param $column
-     * @return mixed
-     * @throws QueryException
-     * @throws \TS\ezDB\Exceptions\ConnectionException
+     * @inheritDoc
      */
-    public function avg($column)
+    public function avg(string $column)
     {
         return $this->aggregate('avg', [$column]);
     }
 
     /**
-     * Find the max of a given column.
-     *
-     * @param $column
-     * @return mixed
-     * @throws QueryException
-     * @throws \TS\ezDB\Exceptions\ConnectionException
+     * @inheritDoc
      */
-    public function max($column)
+    public function max(string $column)
     {
         return $this->aggregate('max', [$column]);
     }
 
     /**
-     *Find the min of a given column.
-     *
-     * @param $column
-     * @return mixed
-     * @throws QueryException
-     * @throws \TS\ezDB\Exceptions\ConnectionException
+     * @inheritDoc
      */
-    public function min($column)
+    public function min(string $column)
     {
         return $this->aggregate('min', [$column]);
     }
 
-    /**
-     * Aggregate Functions. This function clones the current builder and returns the result of function when called.
-     *
-     * TODO: Allow chaining of aggregation methods.
-     *
-     * @param $function
-     * @param string[] $columns
-     * @return mixed
-     * @throws QueryException
-     * @throws \TS\ezDB\Exceptions\ConnectionException
-     */
-    public function aggregate($function, $columns = ['*'])
+    protected function aggregate(string $function, array $columns = ['*']): IAggregateQuery
     {
-        $results = (clone $this)
-            ->setModel(null)
-            ->setBinding(compact('function', 'columns'), 'aggregate')
-            ->get($columns);
-
-        return $results[0]->$function;
+        $parent = $this->clone();
+        $parent->addClause('aggregate', ['function' => strtoupper($function), 'alias' => $function, 'columns' => $columns]);
+        return new AggregateQuery($parent);
     }
 
-
-    /**
-     * @param $operator
-     * @return bool
-     */
-    public function isInvalidOperator($operator)
+    protected function isInvalidOperator(string $operator): bool
     {
         /*
          * isset search is faster than in_array
          * combining isset and for loop is still faster than in_array
          */
         return !isset($this->operators[$operator]);
-    }
-
-    /**
-     * Return current datetime (to be used with mysql)
-     * It returns the current time in PHP's timezone.
-     * Make sure the timezone between the php server and the mysql server match.
-     *
-     * @return string
-     */
-    protected function now()
-    {
-        return date("Y-m-d H:i:s");
     }
 }
