+++ src/Query/Builder/QueryBuilderType.php
@@ -1,4 +1,11 @@
 <?php
+/*
+ * Copyright (c) 2022 - Terminus Studio (https://Terminus.Studio)
+ *
+ * ezDB - https://github.com/TerminusStudio/ezDB
+ *
+ * @license https://github.com/TerminusStudio/ezDB/blob/dev/LICENSE.md (MIT License)
+ */
 
 namespace TS\ezDB\Query\Builder;
